Removed count from explode, to fix proxy downloads.
If downloading through a proxy server, from an HTTPS server, there will be a couple
of CONTINUE sections, which does not get counted by Curl as redirects, and then ends
up as part of the file content.
Since the getResponse() method only uses array_pop() on the resulting array, we can
safely split all secions of the response.
This only means that further redirect and CONTINUE header pairs are left unused in
the array.