Removed count from explode, to fix proxy downloads.
If downloading through a proxy server, from an HTTPS server, there will be a CONNECT
call to the proxy, and a couple of lines in the response from the proxy, which does
not get counted by Curl as redirects, and then ends up as part of the file content.

Since the getResponse() method only uses array_pop() on the resulting array, we can
safely split all secions of the response.

This only means that further redirect and CONNECT header pairs are left unused in
the array.